# [SPARK-32350][CORE] Add batch-write on LevelDB to improve performance of HybridStore