@tgravescs 2020-07-21
I likely won't have time for a review so go ahead without mine