@mridulm 2020-08-10
Sorry for the delay in getting to this.
The only additional change I had made in my version was to batch the batch update.
For example, if the list is very large, the write will block all other threads for a very long time - which has fairness issues; in addition to the serialization consuming too much memory, etc.
Instead, I batched updates to some K values (In my case K was 128).

This can be trivially done with an inner loop doing `for (List<?> batchList : Iterables.partition(entry.getValue(), batchSize)) { ... }`