@baohe-zhang 2020-08-10
This seems an important improvement. Should I put up a followup PR to include this change?